init bajak

+++ database/migrations/1731493356678_create_user_financials_table.ts
@@ -14,7 +14,7 @@ export default class extends BaseSchema {
         .references('id')
         .inTable('users')
         .onDelete('CASCADE')
-      table.bigint('balance').defaultTo(0)
+      table.integer('balance').defaultTo(0)
       table.string('card', 20).nullable().defaultTo(null)
       table.string('sheba', 40).nullable().defaultTo(null)
       table.timestamps()
